backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pyjwt
sqlalchemy
psycopg2-binary
resend
apscheduler
google-genai
gitpython